test(reply): keep directive mocks local to each suite
