+++ src/auto-reply/reply/directive-handling.model-selection.test.ts
@@ -1,17 +1,88 @@
-import { describe, expect, it } from "vitest";
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { loadProviderScopedThinkingCatalog } from "../../agents/model-catalog.runtime.js";
+import { persistStickyModelSelectionBestEffort } from "../../agents/sticky-model-selection.js";
+import type { SessionEntry } from "../../config/sessions.js";
+import { triggerSessionPatchHook } from "../../gateway/session-patch-hooks.js";
+import { enqueueSystemEvent } from "../../infra/system-events.js";
 import { MODEL_SELECTION_LOCKED_MESSAGE } from "../../sessions/model-overrides.js";
-import { lifecycleEvents } from "./directive-handling.mixed-inline.test-harness.js";
+import {
+  onSessionLifecycleEvent,
+  type SessionLifecycleEvent,
+} from "../../sessions/session-lifecycle-events.js";
 import {
   applyMixedDirectives,
   createSessionEntry,
-  enqueueSystemEvent,
-  persistenceMocks,
-  persistStickyModelSelectionBestEffort,
-  refreshQueuedFollowupSession,
-  triggerSessionPatchHook,
-} from "./directive-handling.mixed-inline.test-mocks.js";
+} from "./directive-handling.mixed-inline.test-helpers.js";
+import { refreshQueuedFollowupSession } from "./queue.js";
+
+type PersistenceResult =
+  | { status: "current"; entry: SessionEntry }
+  | { status: "model-selection-locked"; entry: SessionEntry }
+  | { status: "lifecycle-invalidated"; error: string; entry?: SessionEntry };
+
+vi.mock("../../agents/model-catalog.runtime.js", () => ({
+  loadProviderScopedThinkingCatalog: vi.fn(async () => []),
+}));
+
+const persistenceMocks = vi.hoisted(() => ({
+  persist: vi.fn<(params: { entry: SessionEntry }) => Promise<PersistenceResult>>(),
+}));
+
+vi.mock("../../agents/agent-scope.js", () => ({
+  listAgentEntries: vi.fn(() => []),
+  resolveAgentConfig: vi.fn(() => ({})),
+  resolveAgentModelFallbacksOverride: vi.fn(() => undefined),
+  resolveAgentDir: vi.fn(() => "/tmp/agent"),
+  resolveSessionAgentIds: vi.fn(() => ({ requestedAgentId: "main", sessionAgentId: "main" })),
+  resolveSessionAgentId: vi.fn(() => "main"),
+  resolveDefaultAgentId: vi.fn(() => "main"),
+}));
+
+vi.mock("../../agents/sandbox.js", () => ({
+  resolveSandboxRuntimeStatus: vi.fn(() => ({ sandboxed: false })),
+}));
+
+vi.mock("../../agents/sticky-model-selection.js", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("../../agents/sticky-model-selection.js")>()),
+  persistStickyModelSelectionBestEffort: vi.fn(),
+}));
+
+vi.mock("../../gateway/session-patch-hooks.js", () => ({
+  triggerSessionPatchHook: vi.fn(),
+}));
+
+vi.mock("../../infra/system-events.js", () => ({
+  enqueueSystemEvent: vi.fn(),
+}));
+
+vi.mock("./queue.js", () => ({
+  refreshQueuedFollowupSession: vi.fn(),
+}));
+
+vi.mock("./session-entry-persistence.js", () => ({
+  persistReplySessionEntry: (params: { entry: SessionEntry }) => persistenceMocks.persist(params),
+}));
 
 describe("mixed inline directives / model selection", () => {
+  let lifecycleEvents: SessionLifecycleEvent[];
+  let unsubscribeLifecycle: () => void;
+
+  beforeEach(() => {
+    lifecycleEvents = [];
+    unsubscribeLifecycle = onSessionLifecycleEvent((event) => lifecycleEvents.push(event));
+    vi.clearAllMocks();
+    vi.mocked(loadProviderScopedThinkingCatalog).mockReset().mockResolvedValue([]);
+    vi.mocked(persistStickyModelSelectionBestEffort).mockReturnValue("requested");
+    persistenceMocks.persist.mockImplementation(async ({ entry }) => ({
+      status: "current",
+      entry: { ...entry },
+    }));
+  });
+
+  afterEach(() => {
+    unsubscribeLifecycle();
+    vi.restoreAllMocks();
+  });
   describe.each(["", "please reply "])("model scope with prefix %j", (prefix) => {
     it.each([
       { scope: undefined, flag: "", owner: true, target: undefined, writes: false },
